Basic fingering, fledgling API
This is basically me noodling with various techniques for testing the API.

My goals are as follows:

1) clean
2) fast

I think I've managed a good balance, and I've put together a reasonably clean
stack. I'm going to iterate on this and see where it gets me, then start playing
with Kevin's spec as implemented.