Specify the output of the user JSON explicitly.
I'm trying to keep things as simple as possible here. By specifying the output
of the user serializable_hash in the model, it means any time the user's
to_json (or to_xml) is called, it returns the same thing, keeping controllers
nice and free from those kinds of decisions. It might make more sense
to do it in a less explicit way, so as not breaking the API.