A fledgling javascript workflow
If we're treating the API as first-class, then we can go with a pure
javascript workflow.

We could of course consume our own API through Rails, but I think
we can probably assume enough about our users to get away with this.

Again, I've gone for simplicity: History.js to support cross-browser
pushState and fallbacks, with a super, super simple routing layer.

We can improve on this later, I'm sure, but for now I'm quite happy that
it's the simplest thing that works, and that it's really nice and simple.