Sticking with capybara-webkit
I've been trying to find the best balance of testing javascript as part of the
test suite and getting good performance.

capybara-webkit is the best balance of both out there, running in a decent
amount of time, without spinning up a browser. It has it's limitations,
but I'd rather test these manually if possible.