Display the manifest after submitting the initial form.
So, the flow is starting to come together now, and we're getting the basic workflow.

I'm thinking the simple javascript approach is now starting to show it's limitations,
but I do love the simplicity of how Rails serves up the templates.

I'll think about whether we need to involve a framework or not. So far it's pretty clean.