Set up a route for the manifest.
I do feel a little bit like I might be re-inventing the wheel, but I do
still feel like I'm on the right side of simple.

It's always a fine line, I could continue to abstract, or I could
fall back and trust a framework.

This change simply sets up a route for the framework. The main problem is that
it might be a bit too much more effort than it's worth.