Implement support for pushState history.
The main problem here is that there's no hashtag fallback, so this won't work on
older browsers. It's not that big a deal immediately, but again, it could
be an indication that we need to move to a JS framework at some stage.

The key will be to implement a good test suite to test the JS flow, and then
we can experiment with other JS workflows if needed.