This is the first bit of doing things right with Backbone.
Basically, you have your single item submission form, and it's tied to a new
model instance at the point of hitting 'submissions/new' ... this is pretty
powerful, since the weight of the rendering is all down to the state of the
initial submission, so we can do lots of interesting things later.

Very well played, Backbone.js, very well played.