First user flow with Backbone in place.
This is a pretty good milestone, and one that tells me that Backbone
is pretty much exactly what we need on this project.

I'm glad I put the work in to simplify the template compilation, and
I'm personally very happy with mustache.js for client side templating.

It's still early enough to use a different client side templating engine,
but this is enough to prove, to me at least, that Backbone really has the
chops: the features, the simplicity and the power to do all of our
client-side app needs.

The force is strong with this one.