JS client-side flow is complete.
This represents the first complete flow, using Backbone and Rails to seamlessly
keep the view completely client side.

There's a bit of cheating going on with the Rails my/measurements_controller,
but Backbone recommends that any objects that should be populated server-side
should be populated server-side. We're outputting json, so there's no
actual duplication of work. I'm totally ok with that.