Update the view by setting a "saving" attribute on the model
This now completes my ideal workflow, using Backbone, at least from the backend,
by validating the measurement client-side before it's submitted.

Just before the object is saved to the server, the validation runs, and an alert
is displayed if it's invalid. I'm not sure yet how the UI should reflect this,
I suspect that the fields needing to be valid should display themselves as such.

But for now, we have a reference technique (I think), for every basic interaction
that I think it's desirable to have, at least for the submission workflow.

When it comes to collections and things like that, there's more work to be done!