Only ever create one instance of the newMeasurementView
It turns out that in Backbone, if a view takes up the entire UI, you only
want to create it once, to avoid having to re-bind objects to it.

So now, instead of creating new instances of the newMeasurement view, I'm
simply changing the model associated with the existing one and re-initializing

Feels a bit wrong, and we'll have to use the same technique no doubt for the
show view once there are any interactions for it.