Introduce postGIS
PostGIS is a set of functions for PostgreSQL that brings comptability for
geospatial data.

spatial_adapter is a ruby adapter for ActiveRecord that brings spacial-data
support to Rails.

When you combine the two, you get the ease and simplicity of Rails with the
speed and flexibility of true geometric calculations at the database level.

Sounds like a win to me.