fixed delayed_job spec issue
before(:all)ing the work_off was resulting in the thread running after
the scenarios for some reason.  moving the directive to the let! causes
it to get executed in each scenario when assigning updated_result
(caching also seemed to be a problem)

also, added the measurement_count to the updated attributes after the
sql COPY