This is the initial full stack workflow.
The workflow is now full-stack, with rails handling the server side case
of submission, with everything needed to do it completely client side too.

It would be super great if the server side methods consumed the API too, rather
than talked to the database directly : there's some duplication here, but we can get to that later.