Fix API DOCS link in top nav.
I also put back the respond_to :html for the API in order to green the
measurement submission test that broke when I removed it.

I'm going to open an Issue about this in a moment to avoid stepping on
toes, but the original intention of responding to html with the API was
to allow for documentation, not accommodate form submission.

As it is now, it works, but it exhibits some inconsistency at a design
level that's a bit bothersome.